fix: adicionada tratativa de erro em excluirTarefa para tarefa nao localizada

+++ src/main/java/com/marcelo/domain/services/TarefaDomainServiceImpl.java
@@ -4,6 +4,7 @@ import java.time.Instant;
 import java.time.LocalDate;
 import java.time.ZoneId;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
@@ -16,7 +17,6 @@ import com.marcelo.domain.dtos.CriarTarefaRequestDto;
 import com.marcelo.domain.dtos.EditarTarefaRequestDto;
 import com.marcelo.domain.entities.Tarefa;
 import com.marcelo.domain.entities.Usuario;
-import com.marcelo.domain.exceptions.EmailJaCadastradoException;
 import com.marcelo.domain.exceptions.TarefaNaoLocalizadaException;
 import com.marcelo.domain.interfaces.TarefaDomainService;
 import com.marcelo.infrastructure.repositories.TarefaRepository;
@@ -66,6 +66,11 @@ public class TarefaDomainServiceImpl implements TarefaDomainService{
 	@Override
 	public ConsultarTarefaResponseDto excluirTarefa(UUID idTarefa, UUID idUsuario) {
 		
+	    Optional<Tarefa> tarefaOptional = tarefaRepository.findById(idTarefa);
+	    if (!tarefaOptional.isPresent()) {
+	        throw new TarefaNaoLocalizadaException();
+	    }
+		
 		Tarefa tarefa  = tarefaRepository.findByIds(idTarefa, idUsuario);
 		
 		tarefaRepository.delete(tarefa);
@@ -91,7 +96,6 @@ public class TarefaDomainServiceImpl implements TarefaDomainService{
 	@Override
 	public List<ConsultarTarefaResponseDto> consultarTarefas(Instant dataMin, Instant dataMax, UUID idUsuario) {
 		
-		// testar
 		LocalDate localDateMin = dataMin.atZone(ZoneId.systemDefault()).toLocalDate();
 		LocalDate localDateMax = dataMax.atZone(ZoneId.systemDefault()).toLocalDate();
